Reduce event loss that occurs when many scripts fire events at once such as with llResetScript.

+++ Aurora/AuroraDotNetEngine/Helpers/Helpers.cs
@@ -297,7 +297,7 @@ namespace Aurora.ScriptEngine.AuroraDotNetEngine
         //This is the current spot that the event is at in processing
         public EnumeratorInfo CurrentlyAt;
         //This is used to check whether the script has been updated since the last attempt to start
-        public int VersionID;
+        public long VersionID;
         //The times this event has been run through the scheduler
         public int RunningNumber;
         //The currently running state that this event will be fired under
